+++ Gems/LmbrCentral/Code/Source/Shape/CapsuleShape.cpp
@@ -13,7 +13,6 @@
 #include <AzCore/Math/Transform.h>
 #include <AzCore/Serialization/EditContext.h>
 #include <AzCore/Serialization/SerializeContext.h>
-#include <CryCommon/Cry_GeoDistance.h>
 #include <MathConversion.h>
 
 namespace LmbrCentral
@@ -183,12 +182,8 @@ namespace LmbrCentral
         AZStd::shared_lock lock(m_mutex);
         m_intersectionDataCache.UpdateIntersectionParams(m_currentTransform, m_capsuleShapeConfig, &m_mutex);
 
-        const Lineseg lineSeg(
-            AZVec3ToLYVec3(m_intersectionDataCache.m_basePlaneCenterPoint),
-            AZVec3ToLYVec3(m_intersectionDataCache.m_topPlaneCenterPoint));
-
-        float t = 0.0f;
-        float distance = Distance::Point_Lineseg(AZVec3ToLYVec3(point), lineSeg, t);
+        float distanceSq = AZ::Intersect::PointSegmentDistanceSq(point, m_intersectionDataCache.m_basePlaneCenterPoint, m_intersectionDataCache.m_topPlaneCenterPoint);
+        float distance = AZ::Sqrt(distanceSq);
         distance -= m_intersectionDataCache.m_radius;
         const float clampedDistance = AZStd::max(distance, 0.0f);
         return clampedDistance * clampedDistance;
